+++ spine-ts/spine-construct3/ts-defs/editor/IPluginInfo.d.ts
@@ -38,6 +38,7 @@ declare namespace SDK {
 
 		SetIsResizable(v: boolean): void;
 		SetIsRotatable(v: boolean): void;
+		SetIsRotatable3D(v: boolean): void;
 		SetSupportsZElevation(v: boolean): void;
 		SetHasImage(v: boolean): void;
 		SetDefaultImageURL(v: string): void;
@@ -58,6 +59,7 @@ declare namespace SDK {
 		AddCommonPositionACEs(): void;
 		AddCommonSizeACEs(): void;
 		AddCommonAngleACEs(): void;
+		AddCommon3DRotationACEs(): void;
 		AddCommonAppearanceACEs(): void;
 		AddCommonZOrderACEs(): void;
 		AddCommonSceneGraphACEs(): void;