+++ spine-ts/spine-construct3/ts-defs/editor/model/IProjectFile.d.ts
@@ -4,5 +4,6 @@ declare namespace SDK {
 		GetPath(): string;
 		GetProject(): SDK.IProjectFile;
 		GetBlob(): Blob;
+		GetSID(): number;
 	}
 }